+++ package/src/index.ts
@@ -1,5 +1,6 @@
 export { AnnotationCanvas } from "./components/AnnotationCanvas";
 export type { CanonicalAnnotation, LabelMap, SymbolSize, SymbolSizeUnit, ToolType } from "./types/canonical";
+export type { CommentAnchor, CommentDraft, CommentTarget, CommentUndoOp } from "./types/comments";
 export type { ThemeVars } from "./theme";
 export { DEFAULT_THEME } from "./theme";
 export { geo } from "./utils/geometry";